LayoutTests/inspector/console-tests-expected.txt

CONSOLE MESSAGE: line 8: log
CONSOLE MESSAGE: line 9: debug
CONSOLE MESSAGE: line 10: info
CONSOLE MESSAGE: line 11: warn
CONSOLE MESSAGE: line 12: error
CONSOLE MESSAGE: line 13: Message format number %i, %d and %f
CONSOLE MESSAGE: line 14: Message %s for %s
CONSOLE MESSAGE: line 15: Object %o
CONSOLE MESSAGE: line 17: repeated
CONSOLE MESSAGE: line 17: repeated
CONSOLE MESSAGE: line 17: repeated
CONSOLE MESSAGE: line 17: repeated
CONSOLE MESSAGE: line 17: repeated
CONSOLE MESSAGE: line 24: 1
CONSOLE MESSAGE: line 25: [object HTMLDocument]
CONSOLE MESSAGE: line 26: [object HTMLHtmlElement]
Tests that console logging dumps proper messages.

console-tests.html:8log console-message console-js-source console-log-level
console-tests.html:9debug console-message console-js-source console-log-level
console-tests.html:10info console-message console-js-source console-log-level
console-tests.html:11warn console-message console-js-source console-warning-level
console-tests.html:12error console-message console-js-source console-error-level
console-tests.html:13Message format number 1, 2 and 3.5 console-message console-js-source console-log-level
console-tests.html:14Message format for string console-message console-js-source console-log-level
console-tests.html:15Object Object console-message console-js-source console-log-level
5console-tests.html:17repeated console-message console-js-source console-log-level repeated-message
console-tests.html:19count: 1 console-message console-js-source console-log-level
console-tests.html:19count: 2 console-message console-js-source console-log-level
console-tests.html:20group console-message console-js-source console-log-level console-group-title
console-message console-js-source console-log-level
console-tests.html:23timer: 0ms console-message console-js-source console-log-level
console-tests.html:241 2 3 console-message console-js-source console-log-level
console-tests.html:25HTMLDocument console-message console-js-source console-log-level
console-tests.html:26 console-message console-js-source console-log-level
